pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Q&A Backend Service"
authors = ["Freedom Loisa <kitakayaloisa@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.1.3"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
django-ninja = "^1.3.0"
django-cors-headers = "^4.6.0"
whitenoise = "^6.8.2"
gunicorn = "^23.0.0"
pyjwt = "^2.10.1"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
